moved transformer to separate class
